Add paginated table in proifle page with 3 columns instead of 2
The new table shows unique identites that are searched before
by the input "Search unique identities". The new profile page
will show 3 columns, first is the basic info of the profile and if
you want to add more identities/orgs to the progfile the second
column will appear and will show the list of organizations or the
input "Search unique identites" described before.

In the third column will be the info of the unique identity
clicked from the unique identites table.