[hatstall] Add support for editing organizations and domains
Now, every entry of the "Organizations" table in that view
will show a list of domains for that organization and a "Edit"
button. When clicking this button, a new view will be rendered
loading a form which allows to:
* Remove the organization (and its domains) from the database.
* Remove its domains individually.
* Add a new domain.

When the "Remove organization" button is clicked, a modal will
reveal asking the user for confirmation. Once the "confirm" button
is clicked, the form will be submitted and that organization and
its domains will be deleted from the database.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>