Support assign same organization in different time frames
When an enrollment is first added, Hatstall uses the default min
and max dates (date range `1900-01-01` to `2100-01-01`). When an
enrollment was set to be updated, the method deleted it first
in Sortinghat before adding the one with the new dates. By doing
this, the method was deleting also other enrollments for the
given organization with other date ranges (if any), which were
always included within the default date range.

After this change, adding different time frames for a given
organization is supported by using the `Enroll` SortingHat
command, with the `merge` option set to `True`. This means
overlapping date ranges for a given organization will be merged.

To initialize the object from this command's class is necessary
to provide the database connection parameters, which are taken
from the `Conf` class in Hatstall. The parameter `port` has been
added to this class with a default value, similar to other methods
using the `Database` class which has the same default port value.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>